dashboard/dashboard_requirements.txt
pandas
python-dotenv
awswrangler
streamlit
plotly
